Make app sync with DB (#4)
* add: Passing BoardId btw navGraphs

* fix: Update Card details functionality

* add: Dialog for updating the List title.

* fix: Create Card Flow with selected boardId and Listid

* random Images Cover for boards.

+++ core/data/src/main/java/com/jetapps/jettaskboard/local/entity/ListEntity.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 )
 data class ListEntity(
     @PrimaryKey(autoGenerate = true)
-    val listId: Int = 0,
+    val listId: Long = 0,
     val title: String,
-    val boardId: Int
+    val boardId: Long
 )
